@@ -38,14 +38,12 @@
                             "game": {
                                 "type": "string",
                                 "enum": [
-                                    "fallout4"
+                                    "fallout4",
+                                    "skyrimSpecialEdition"
                                 ]
                             },
                             "projectPath": {
                                 "type": "string"
-                            },
-                            "port": {
-                                "type": "integer"
                             }
                         },
                         "required": [
@@ -207,8 +205,9 @@
         "fast-deep-equal": "^2.0.1",
         "ini": "^1.3.5",
         "md5-file": "^4.0.0",
-        "rxjs": "^6.4.0",
-        "vscode-debugprotocol": "^1.34.0",
+        "ps-list": "^6.3.0",
+        "rxjs": "^6.5.2",
+        "vscode-debugprotocol": "^1.35.0",
         "vscode-languageclient": "^5.2.1",
         "winreg": "^1.2.4",
         "xml-writer": "^1.7.0"
@@ -218,17 +217,17 @@
         "@semantic-release/changelog": "^3.0.2",
         "@semantic-release/git": "^7.0.8",
         "@types/ini": "^1.3.30",
-        "@types/mocha": "^5.2.6",
-        "@types/node": "^11.11.3",
+        "@types/mocha": "^5.2.7",
+        "@types/node": "^12.0.4",
         "@types/winreg": "^1.2.30",
         "cpr": "^3.0.1",
-        "prettier": "^1.16.4",
+        "prettier": "^1.17.1",
         "rimraf": "^2.6.3",
-        "semantic-release": "^15.13.3",
-        "semantic-release-vsce": "^2.2.6",
-        "tslint": "^5.14.0",
-        "typescript": "^3.3.3333",
-        "vscode": "^1.1.30"
+        "semantic-release": "^15.13.12",
+        "semantic-release-vsce": "^2.2.8",
+        "tslint": "^5.17.0",
+        "typescript": "^3.5.1",
+        "vscode": "^1.1.34"
     },
     "displayName": "Papyrus",
     "engines": {
